refactor: contain cpal behind an internal AudioBackend seam
Introduce a pub(crate) AudioBackend trait as the single place the crate
reaches the platform audio library, with CpalBackend as its sole
implementation: device enumeration, selector resolution, and input/output
stream opening now go through the trait, and cpal types appear only inside
the cpal implementation and the small decibri-owned BackendDevice /
BackendStream wrappers. device.rs keeps the public device types and the pure
is_default helper and delegates listing through the seam; microphone.rs and
speaker.rs hold a BackendDevice and a BackendStream and open through the
backend, passing the realtime capture/render and error callbacks unchanged.

Behaviour-preserving: open still accepts or fails at the requested config,
stop() still drops the held stream to release the device, the stream types
stay Send + Sync, and the typed DeviceFailed path through take_last_error()
is intact. No public API change; the seam is internal.
